+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/port/LoanMapperPort.java
@@ -1,7 +1,7 @@
 package com.asc.loanservice.domain.loan.application.port;
 
-import com.asc.loanservice.contracts.LoanApplicationCreateInput;
-import com.asc.loanservice.contracts.LoanApplicationRequest;
+import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationCreateInput;
+import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationRequest;
 
 public interface LoanMapperPort {
     LoanApplicationCreateInput createInput(LoanApplicationRequest loanApplicationRequest);
