+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/LoanEvaluatorProviderAdapter.java
@@ -1,6 +1,6 @@
 package com.asc.loanservice.infrastructure;
 
-import com.asc.loanservice.contracts.LoanApplicationRequest;
+import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationRequest;
 import com.asc.loanservice.domain.loan.application.port.LoanEvaluatorProviderPort;
 import com.asc.loanservice.infrastructure.rule.Rule;
 import org.springframework.stereotype.Service;
