+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/api/contracts/LoanApplicationEvaluationStatus.java
@@ -1,4 +1,4 @@
-package com.asc.loanservice.contracts;
+package com.asc.loanservice.infrastructure.api.contracts;
 
 public enum LoanApplicationEvaluationStatus {
     APPROVED,
