+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/api/contracts/LoanApplicationResult.java
@@ -1,4 +1,4 @@
-package com.asc.loanservice.contracts;
+package com.asc.loanservice.infrastructure.api.contracts;
 
 import lombok.*;
