+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/repository/model/LoanApplication.java
@@ -2,8 +2,8 @@ package com.asc.loanservice.infrastructure.repository.model;
 
 
 
-import com.asc.loanservice.contracts.LoanApplicationEvaluationStatus;
-import com.asc.loanservice.contracts.LoanApplicationView;
+import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationEvaluationStatus;
+import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationView;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
