+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/rule/Rule.java
@@ -1,6 +1,6 @@
 package com.asc.loanservice.infrastructure.rule;
 
-import com.asc.loanservice.contracts.LoanApplicationRequest;
+import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationRequest;
 
 public interface Rule {
     Boolean isValid(LoanApplicationRequest loanApplicationRequest);
